Move operations to its own subpackage
