+++ megaradrp/processing/__init__.py
@@ -1,0 +1,4 @@
+
+from .trimover import OverscanCorrector, TrimImage
+from .aperture import ApertureExtractor, ApertureExtractor2
+from .fiberflat import FiberFlatCorrector
